requirements.txt
mkdocs-material
PyMuPDF
Pillow
PyYAML
